requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai>=1.30.0
aiohttp==3.9.1
pytz
azure-search-documents==11.4.0
azure-ai-vision-imageanalysis
azure-cognitiveservices-vision-computervision
beautifulsoup4
lxml
python-multipart
azure-identity==1.18.0
pydub==0.25.1
azure-cognitiveservices-speech==1.40.0
azure-keyvault-secrets==4.7.0
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
